# Powradix optimizations