--- a/src/electionguard/discrete_log.py
+++ b/src/electionguard/discrete_log.py
@@ -4,9 +4,8 @@
 import asyncio
 from typing import Dict, Tuple
 
-from .constants import get_generator
 from .singleton import Singleton
-from .group import ElementModP, ONE_MOD_P, mult_p
+from .group import ElementModP, ONE_MOD_P, mult_p, get_generator_element
 
 DLOG_CACHE = Dict[ElementModP, int]
 DLOG_MAX = 100_000_000
@@ -73,7 +72,7 @@ def compute_discrete_log_cache(
     max_element = list(cache)[-1]
     exponent = cache[max_element]
 
-    g = ElementModP(get_generator(), False)
+    g = get_generator_element()
     while element != max_element:
         exponent = exponent + 1
         if exponent > DLOG_MAX: