[//]: # (🚨 Please review the CONTRIBUTING.md in this repository. 💔Thank you!)

### Issue
Fixes #410 

### Description
This is a second draft of a fix for #410, addressing hopefully most of the comments in the first one. A few things of note:
- There was a discussion between Olivier Pereira and Josh Benaloh about some of the checks that happen in the disjunctive Chaum-Pedersen proof verifier as being redundant. Fixing this yielded a 40% speedup in proof verification.
- `get_constants()` was calling `getenv()` enough to have a significant impact on performance. The fix is that the `ElectionConstants` structure is now computed once and cached. 20% speedup, just from that.
- We also benefit from only keeping a single instance of the `xmpz` values for P, Q, and G. This PR includes a second data structure, `ElectionConstantInternals`, which holds those and other related values.
- The original code used an environment variable, `PRIME_OPTION`, to select between test primes and standard primes. I've added `POWRADIX_OPTION` to similarly specify the precompute memory use.
- We need to be able to have the unit test runner, on GitHub, set `PRIME_OPTION` one way, while having unit tests that explicitly work with the large primes. If you look in `constants.py`, I've added a stack-like structure, with pushing and popping operations, that manage all of the relevant constants. This avoids unit tests causing side-effects that are observable by other unit tests, while still having all the benefits of the caching. 
- I rethought the type hierarchy for `BaseElement`, `ElementModP`, and the new `ElementModPAcceleratedPow`. The aforementioned environment issue got rid of the need for a configuration parameter to the `ElementModPAcceleratedPow` constructor. Also, so far as I can tell, the bounds check in the `BaseElement` constructor served no purpose, since when we really care we use the `is_in_bounds()` method. These observations together allowed me to get rid of all the configuration parameters for all of the constructors, which in turn simplified everything, since Python inheritance with custom `__new__` and `__init__` methods is very messy.

### Testing
The full unit test suite passes when `PRIME_OPTION` is set to `TestOnly`. If you don't set that environment variable, then a bunch of the unit tests bomb out with Hypothesis time limits. I didn't want to change all those time limits, but it's a thing worth considering.

### Performance improvements
Here's the result of running the benchmark:
```
Benchmark: PowRadixOption.NO_ACCELERATION, problem size: 500
  Precompute time: 0.000 sec
  Creating Chaum-Pedersen proofs:     8.242 +/- 0.406 ms / proof
  Validating Chaum-Pedersen proofs:  17.573 +/- 0.723 ms / proof
Benchmark: PowRadixOption.LOW_MEMORY_USE, problem size: 500
  Precompute time: 0.125 sec
  Creating Chaum-Pedersen proofs:     1.613 +/- 0.147 ms / proof
  Validating Chaum-Pedersen proofs:  11.065 +/- 0.599 ms / proof
Benchmark: PowRadixOption.HIGH_MEMORY_USE, problem size: 500
  Precompute time: 2.393 sec
  Creating Chaum-Pedersen proofs:     1.103 +/- 0.053 ms / proof
  Validating Chaum-Pedersen proofs:  10.350 +/- 0.274 ms / proof
Benchmark: PowRadixOption.EXTREME_MEMORY_USE, problem size: 500
  Precompute time: 15.403 sec
  Creating Chaum-Pedersen proofs:     0.958 +/- 0.049 ms / proof
  Validating Chaum-Pedersen proofs:  10.212 +/- 0.288 ms / proof
Speedup of PowRadixOption.LOW_MEMORY_USE    for proofs      :  5.109
Speedup of PowRadixOption.LOW_MEMORY_USE    for verification:  1.588
Speedup of PowRadixOption.HIGH_MEMORY_USE   for proofs      :  7.472
Speedup of PowRadixOption.HIGH_MEMORY_USE   for verification:  1.698
Speedup of PowRadixOption.EXTREME_MEMORY_USE for proofs      :  8.605
Speedup of PowRadixOption.EXTREME_MEMORY_USE for verification:  1.721
```