@danwallach 2021-11-05
So are you ready for this as a final PR?